fix(Calculation) : change double to integer

+++ src/main/java/calculator/Application.java
@@ -16,7 +16,7 @@ public class Application {
             String separator = validation.validateString(inputString);
 
             Calculation calculation = new Calculation();
-            Double result = calculation.calculationString(separator,inputString);
+            Integer result = calculation.calculationString(separator,inputString);
             System.out.println("결과 : " + result);
         }catch (IllegalStateException err){
             return;
